backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "livekit-agents[google,images,silero]~=1.2",
    "livekit-plugins-bey>=1.3.12",
    "livekit-plugins-noise-cancellation~=0.2",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
